requirements.txt
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
apprise==1.9.4
asgi-lifespan==2.1.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
coolname==2.2.0
cryptography==45.0.7
cycler==0.12.1
databricks-sdk==0.65.0
dateparser==1.2.2
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.116.1
Flask==3.1.2
fonttools==4.59.2
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
griffe==1.14.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lightgbm==4.6.0
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
numpy==2.2.6
nvidia-nccl-cu12==2.28.3
oauthlib==3.3.1
omegaconf==2.3.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pendulum==3.1.0
pillow==11.3.0
prefect==3.4.17
prometheus_client==0.22.1
protobuf==6.32.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
python-socks==2.7.2
pytz==2025.2
PyYAML==6.0.2
readchar==4.2.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.3
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.8.15
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
xgboost==3.0.5
zipp==3.23.0
